Make jar creation work
Creating the jar was not possible because the code was trying to
read a resource and use it as a File object for the PropertiesUserManagerFactory
setFile method (Apache ftpserver).

ftpserver already managed the case when it had to read the resource from
a jar file, so I shouldn't have to create the File from the resource myself,
but just create an abstract File object with the relative path of the resource.

That was enough for ftpserver to find the resource in the Jar file.

Check the loadFromFile method of PropertiesUserManager class of Apache
ftpserver to see how they do it (using getResourceAsStream).